Several changes to build structure - contact Dr Mike if you hit problems.
I've made most of the Makefile's use the Makefile.inc include in the
root of the anaconda src tree to get the python path, instead of having
it defined in many different places.

In the top level Makefile I added a install target called install-unconfig.
This target installs the anaconda script and associated support files
with the intention of being able to run anaconda in unconfig mode after
a successful install.  This mode is automatically selected if
the executed script is called 'anaconda-unconfig', or the '-U' command
line parameter is used.