register a parted exception handler for gui mode
modified isys.makeDevInode to use /dev/foo if it is there, while
preserving backwards compatible behavior

import fsset for workstation, it needs to find the ext2 filesystemtype
for some reason

skip drives that can't be opened or partition tables that can't be
read.  we still need to handle 'fresh' disks better.