we do turnOnFilesystems during the install before migrateFilesystems. since we've switched to mounting ext3 filesystems as ext3 instead of as ext2 (which was the 7.2 behavior), this means we try to mount the filesystem as ext3 before it's been migrated.
ick.  so, if the entry is a) being migrated and b) has a valid origfsystem
we should mount as the origfsystem instead of fsystem (which is the final
post-migration fs) so we change less.

test cases we need to run with this in addition to the standard stuff
the robot would catch:
1) upgrade with migrate
2) install with migrate
3) upgrade with no migrate
4) install with some vfat or reiser partitions