ow.
keep a copy of whether metapkgs and optional components are selected by
default.  then use this information when writing out the anaconda-ks.cfg
so that we can write out something which is correct (#73195, #80838).

note that for metapkgs, we have to write out -pkg for each package in the
group because we have no syntax for unselecting groups and staying that
way seems saner