All hail the hypnotoad!
Yeah, so we have some netlink_* functions now.  You can build nltest in the
isys directory and it'll dump info on your interfaces including the index
number, the name, the MAC address, the IPv4 address, and the IPv6 address.

Next up is rewriting isys functions to use the netlink_* functions.