HERE COMES YOUR WATER! ....
Ok, so what exactly are we looking at here?  A slapped-together pile of stuff that uses libdhcp instead of pump.  I call it a work in progress.  Right now we have support for regular IPv4-based installs.  IPv6 is sort-of halfway there.  I'm still wedging it in to place (i.e., dodging sigsegvs).  I wanted to get an IPv4-libdhcp-move committed now so that it's in CVS.  The pump-compatible API in libdhcp is mostly useless, so I'm moving things from using the core-dumping pump API functions to just using the libdhcp stuff directly.  I had to back out a lot of IPv6 stuff I did in the interim.

Please avoid making huge changes to the loader2 network code because I'm still tearing through that.