Hooray for 4th grade C programming class!
I had an idea that we could make loader smaller if we didn't call toupper() or string functions.  Well, we use string functions all over the place, but I did remove the toupper() calls I used.  Yeah, not exciting.  Believe it or not, this was involved in a large tracking-down of double free()'s yesterday.