Stupid fix for making sure /lib/ld-linux.so.* ends up in stage2.
The DSO deps are different depending on if /lib/ld-2.*.so or /lib/ld-linux.so.*
is called.  In the former case, we don't end up with the symlink so nothing
can run under stage2.  This fix makes sure that symlink gets pulled in, but
it does it in a stupid way.  Feel free to commit something much better later.