Make split media image installs work.
Rework image.py into a support file with no InstallMethod subclasses in it.
Move the media switching code into a yum callback and make sure our base
repos have the correct attribute set to trigger using this callback.  Be more
careful to only attempt ejecting a CD when we're using a method that has a
CD device.