Support SHA256/SHA512 password encoding from kickstart (#427388)
Allow users to select the SHA256 or SHA512 encoding algorithm in
addition to MD5 for password encoding.  This can be accomplished
through kickstart only for RHEL 5, but RHEL 6 will have UI changes
to accomodate this change as well.

Use the authconfig statement in kickstart to set the encoding
algorithm:

    authconfig --enableshadow --enablemd5
    authconfig --enableshadow --passalgo=sha256
    authconfig --enableshadow --passalgo=sha512

Passwords specified with the 'rootpw' command or user accounts with
the 'user' command will be encoded using the specified algorithm.
The /etc/libuser.conf file on the target system will have its
crypt_style setting updated to the preferred algorithm as well.

MD5 encoded passwords start with $1$, SHA256 passwords begin with $5$,
and SHA512 passwords begin with $6$.

Setting the root password through the graphical or text mode interface
to anaconda will still get you an MD5 encoded password.