Use libnl to get interface MAC and IP addresses
Reduce nl.c to enough to communicate with libnl.  We already have
libnl, so we might as well use it.  Plus, the latest version has
a reasonable API, so I'm totally against using it.

The _isys module and loader have been updated to call the new
functions:  nl_mac2str() and nl_ip2str() to get the MAC address or
IP address for the specified interface.

This patch also drops our need for glib for isys.  Yeah, I added the
dep and I'm removing it.