fsset.py tweaks for ext4dev & xfs
ext4dev just got a "test_fs" flag which soon must be set in order for
the kernel to mount it; this will prevent ext3 from accidentally getting
mounted as ext4, and making it hard to go back.  (when ext4 goes prime-time,
the need for this flag will be removed).

Also tweak the xfs class a bit; most mkfs arguments specified are actually
defaults now, and there is a nicer labeling command available, xfs_admin - a
665-byte bash script which I added to upd-instroot too.