Remove old IP addresses from interface on reconfig (#218273)
Remove existing IP addresses on the specified interface during a
reconfiguration of the interface.  The existing code just added
more addresses each time.  The new address would be there, but the
existing one would still be there leading the user to believe the
address never changed (you can compare the output of 'ifconfig'
and 'ip addr') and see that the ip command shows all addresses.

This patch adds a new function called clearInterface() which wraps
our pumpDisableInterface() call.  It uses libnl to find and delete
the existing IP addresses on that interface, then calls the
pumpDisableInterface() function.

Also fixed a SIGSEGV that can occur in STEP_IP in doLoaderMain()
if loaderData->ip is NULL when the check for "query" is done.