Allow creation of LUKSDevice with no passphrase.
The hack in getScheme for autopart causes a traceback if a LUKSdevice is created without a passphrase, which is something that makes sense to do when discovering existing devices. So we just use the "scheme" data member internally as a workaround.

Also, make sure all return values from openDevice are numeric.