Propagate hostname from stage 1 to stage 2 on s390x (#354021).
The hostname and DNS settings on s390x are provided by a kickstart
file or by a CMS conf file or by prompting the user in stage 1.
Either way, we already have all of that information before we move
on to stage 2.  The network device post configuration screen in stage
2 was missing the hostname.  This patch fixes that as well as fixes
some misbehavior with reading those settings from /tmp/netinfo in
loader.