Update the method string handling for NFS and URL installs.
yum doesn't understand nfs: and nfsiso: URLs, so we've been mangling them
early in the anaconda file.  Now with stage2=, we need to do this later
because in the NFS case, we won't have the source repo mounted.  Delay the
mangling until we're doing repo setup, which is also where the NFS source
will get mounted.