Focus root password entry box (#436885).
Make sure the root password entry box has the focus when we get to
the root password screen.  Remove the reference to root-password.png
in the glade file to avoid libglade warnings on the console.  Try to
see if we can get Enter keypresses from the confirm password entry
box to trigger the Next button.  It's not working for me, but my
keyboard is acting up with the latest kernel.