Support encrypted RAID member devices. (#429600)
When creating the RAIDDevice instance, pass in Device instances (as
opposed to device names) for the members list. This way we have all
the information we need to set up the RAID, including encryption
stuff. Previously we instantiated Devices for the RAID members in the
setupDevice method. Now we just use the preexisting Device instances
from the member RequestSpecs.