Stop neutering DRI
The attached patch makes anaconda stop neutering DRI/DRM (at least it should.)
It does it by including the DRI userspace modules (from mesa) and the kernel
DRM drivers.

Why do this?

To be more like the 'normal' installed system. Heck, if DRI's going to fail,
the system's just going to blow up on the first boot anyway. Furthermore,
there are cards/chips that do 2D accel via the 3D pipeline.

Plus, if we want to actually disable it in anaconda, we should do it
explicitly, not by accident as a consequence of some files not being
there.

Caveats:
- listing the kernel drivers explicitly is a hack
- this doesn't pull libGL onto the second stage. Assuming nothing dlopens()
  it, that shouldn't be a problem