Make isys.dhcpNetDevice() work in rescue mode (#435978).
Remove the isys.pumpNetDevice() function since it's just a
passthrough for dhcpNetDevice() at this point.  Update the
function in isys.c to take in None for the class name.  Fix
the anyUsingDHCP() function in network.py for rescue mode
and update calls as appropriate.  Update rescue.py to call
dhcpNetDevice() instead of pumpNetDevice().  And update the
list of files for the rescue image to include /sbin/dhclient
and /sbin/dhcp6c and fallbacks for users.