Resizing lvs on top of RAID fails, make the error not a traceback
Resizing an LV on top of raid fails as we don't have the RAID on at the
point we do the metaresizing (#441706).  Changing that around is kind of
scary at this point and it's a pretty rare case, so just avoid
a traceback for now.  And really, we should catch the error anyway.