Write IPv6 values to /etc/sysconfig/... correctly (#433290)
Handle cases where the default gateway is an IPv6 address,
which has to be written as IPV6_DEFAULTGW rather than just
GATEWAY.  The UI still needs work as it is possible to have
a default gateway for IPv4 and IPv6 concurrently, but we
are only prompting for one address.

This patch also removes the duplicate anyUsingDHCP() from
iw/network_gui.py and uses the one in network.py.  It also
adds an anyUsingStatic() function.  Some functions in
network.py have been changed to return True/False rather
than 1/0.

The network configuration screen in anaconda (not stage 1,
but in stage 2) will activate the gateway and DNS entry
fields now if any interface is set to have any static
configuration.  Previously, it would only activate if all
devices were set to manual configuration for everything.
In addition, you can currently leave the fields blank,
which works well for situations such as IPv4 being DHCP
and IPv6 being static with no IPv6 gateway.  The system
will get a default gateway for IPv4 using DHCP in that
case.