Changes related to BZ #230949
Require the latest libdhcp so that 230949 works.  We allowed
users to only enter a single network stack's information when
doing manual configuration, but the library didn't like that.

Also, I had some cleanup code in net.c that I left throughout
the entire process of working on 230949 and since it now works,
I'd rather just leave it here than back it out.  It's all safe.