Change mount point for CD to /mnt/stage2 when looking for stage2 (#443755).
This only affects installations where you are booting the full DVD but (for
some crazy reason) want to do an HD or NFS install.  Otherwise we'll already
have /mnt/source mounted for the CD and won't be able to mount the real
installation media on top of it.  The DVD will end up getting used even though
that's not what you wanted.