Once we've found the stage2 media on CD, always use it (#443736).
What happens now is that we never unmount the stage2 image if the NFS repo is
incorrect, so trying to mount it again fails and we fall through to a whole
other code path.  It's easier to just not try to remount the CD than to do
all the unmounts.