Handle the DVD having a disknumber of ALL. (443291)
Anyway, right now the .discinfo file that is on the DVD has "ALL" in it.
This just comes from the normal tree.  When anaconda is looking for info
about the iso images or discs it gets it looks at the .discinfo file to
see what number it is.  We used to munge the DVD and put in
1,2,3,4,etc.. for how many discs there are.  This kind of sucks, 1)
munging files blows.  2) requires you to know ahead of DVD creation time
how many split isos it would equal.  Pungi code got rearranged to where
2 isn't a trivial fix right now, and I realized that we could probably
just fake anaconda into thinking that "ALL" == 1, as it's essentially
the first of a single DVD sized iso set.