Preserve 'set the hostname' setting when going Next/Back (#443414)
Better fix for #408921.  Using network.overrideDHCPhostname to set
the default behavior of requiring users to set a hostname.  If the
user changed that to 'use DHCP', clicked Next, then clicked Back,
the network configuration screen would be back at manual for the
hostname.  Fixed that by honoring overrideDHCPhostname.  Also took
the opportunity to make overrideDHCPhostname a boolean since that's
how we are using it.