Add support for attaching gdbserver to the loader early on.
If gdbserver is present in /usr/bin/ and the gdb= parameter is given with a
host:port format, start gdbserver up and attach it to the loader process.  This
enables remote debugging of the loader with minimal changes to the initrd.
The easiest way to do this will be to add a second initrd to the PXE configs
or wherever else containing the gdbserver binary and perhaps extra libraries
to make it work.