Check return value of asprintf() consistently
For instances where the asprintf() was ignored, wrap it in
a test to check if it failed.  If it failed, log a message
and abort.  There may be some instances where abort may
not be what we want to do, however, we were ignoring the
return value completely so if we did get a failure, it would
just SIGSEGV.  Now it will SIGABRT instead.