Look for repo config files in /etc/anaconda.repos.d.
yum wants to look in installroot+/etc/yum.repos.d for config files, which
means it ends up looking under /mnt/sysimage.  On fresh installs this is
no problem because that path does not exist, but on upgrades it does.  We
end up with the wrong repo config objects in the upgrade case.  The least
hacky fix is to change where we store repo config files to a path that
likely won't exist on installed systems.