Fix loadkeys on serial console
when running anaconda (kickstart) over a serial linedisplay (two line
output device) loadkeys does not function. The output of anaconda is
writing to /dev/console, which is achieved by kernel parameter
"console=ttyS1". If user input is needed (ks.cfg-%pre) the layout of the
keyboard does not match the settings in ks.cfg. Obviously the loadkeys
of anaconda does not function correctly if the output is a serial port.
This patch fixes this by not using /dev/console, but /dev/tty0 (the
master-tty).