The actual size of a logical volume must be rounted down. (#415871)
the lvm tools automatically round up a logcial volume when it is
being created with a size that is not a multiple of the volume
group physical extent size.  This patch makes sure that when
the logical volume partition request is asked about the actual
size of a logical volume, it returns clamped (rounded) size.