Support xdriver= again
For now, we need to be able to support xdriver= as sometimes, drivers are
broken and people want vesa instead.  This is kind of lame and it'd be
nice if we didn't have to carry such hacks.

So take the argument and write out a minimal xorg.conf if it's given.
This is all hidden away in the anaconda class, although arguably it could
belong in instdata instead.  But if it's in instdata, more things might
actually try to depend or care about it.  And we don't want that.