Support booting from FCP-attached CD/DVD drive on s390 (#184648)
The s390 can boot El Torito CD or DVD images iff they are attached
by zFCP, used as the IPL device, and contain a specially formatted
boot image on the disc.  IBM provided the tool to combobulate the
boot image together and a description of the desired execution
path.

When you boot on s390, the linuxrc.s390 will look to see if you
IPL'ed from a CD or DVD.  If you did, it will ask if you also
want to install from that device.  If you answer yes (y, Y, or
any case spelling of 'yes'), the script will bring the IPL device
online so the kernel assigns it a device name.  Then it skips over
the network configuration and starts you in to loader.  If you
tell it no or did not IPL from a CD or DVD, it'll launch the
missiles--wait, no, I mean you get the normal network installation
questions before loader starts.

I have no way to test this as it requires the following changes:
(1) Rel-eng needs to build s390x media with -no-emul-boot and
    specify the new cdboot.img file on that platform.  I have
    already contacted rel-eng about making this change.
(2) I don't have a CD-ROM drive in my mainframe.  IBM does and
    testing is all falling on them.  IBM knows this...maybe.

I explain all of this like anyone else on the team will ever get a
chance to experience it.

So there you have it.  A letter opener.