Add support for filing tracebacks directly into bugzilla.
This patch adds support for save to bugzilla, using the python-bugzilla module.
We get the bugzilla URL from product.bugUrl and require the user to already
have a valid account with that bugzilla instance.  That should cut down on
potential abuse.

To cut down on the number of possible duplicates, we hash the file name,
function name, and line of each frame in the traceback and store that hash in
the bug itself.  Before filing a new bug, we query for any bugs containing that
hash value.  If found, we simply add the traceback as a new attachment and put
the user on the CC list.  If not found, we create a new bug.  Either way, the
user is encouraged to visit their bug and make a more meaningful comment.