Add the bug filing abstraction layer to anaconda.
This may not be the best final resting place for this code, but I don't
expect there to be any users outside of anaconda for now so it's not worth
the extra complication of findig a source control home and new package for
a single source file.  This could all change in the future, of course.