Support VDSK devices on s390x (#264061)
A VDSK ("virtual disk", I think) on s390x is like a ramdisk
on other platforms.  The kernel sees VDSK devices and gives
them a /dev/dasd* device name.  The 'discipline' of the
device is set to FBA rather than ECKD.  In anaconda when we
would see these VDSK devices, we assumed they needed a
dasdfmt like ECKD devices.  This patch corrects the behavior
and labels VDSK devices as msdos volumes so that we can use
them in anaconda.