mk-images: replace kernel-xen with pv_ops kernel
kernel-xen is dead, replaced now with pv_ops Xen guest
support in the stock kernels.

On x86_64 and ia64, we now just point the [images-xen]
.treeinfo stanza at the stock kernels in images/pxeboot.

On x86_32, things are a bit more complicated. The stock
kernel is kernel.i586 but Xen guests support is only
included in kernel-PAE.i686 because Xen only runs in PAE
mode.

In this case, we just copy the kernel-PAE vmlinuz to
images/pxeboot/vmlinuz-PAE and build another initrd as
images/pxeboot/initrd-PAE.img. It's these images we point
the [images-xen] stanza at.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>