yuminstall: don't ever stop people installing the virt group
This hack was important to stop people installing Xen dom0
on a machine where it wouldn't boot.

We don't have dom0 currently and now always makes sense to
allow people to install virt-manager and co. for containers,
qemu etc. etc.

If the hack ever comes back, it should just be to delete
the xen-hypervisor package (not in comps for F10) from the
list of options.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>