Correctly handle typos in the stage2 location when inferred from repo=.
When we are given repo= but have to infer a location for the stage2 image,
we need to deal with typos.  When we try to get the stage2 image, the user
has a chance to correct things in the UI.  However, this does not change
the repo= parameter we used to make our stage2 guess.  While we can't be
sure what the user really meant to do, we can at least try to trim off the
/images/whatever portion of the path that was constructed by the UI.