Use NetworkManager instead of libdhcp. (#458183)
Finally, no more libdhcp.

This is the first set of changes to take anaconda over to
the wonderful world of NetworkManager.  We are no longer
linking with libdhcp to do interface configuration.  NM
is started early in the installation and opens the door
to things like WPA installation support and things like
that.