Use yum's handling of optional/default/mandatory package selection (#448172).
In kickstart, groups are processed after packages.  This means if you add in a
package that's default for a group, then later go back and add the group without
defaults, the package you specified will not get installed.  This is because we
were processing --nodefaults by removing the packages from the transaction
entirely.  Instead, use yum's builtin group package selection handling which
doesn't do this add/remove stuff.