Select packages after repos are set up (#457583)
Set up which packages are actually selected after you've done your repo
configuration so that we don't have packages set to be installed from repos
which no longer exist.

To do this, we switch to using whether or not the group is default rather
than selecting/deselecting groups in task selection.  Then, move the base
package selecting to be later