Use UUIDs instead of device nodes in crypttab. (#461702)
We're also changing the mapping names to be of the form "luks-$UUID".
Additionally, we use device nodes for LUKS devices in the fstab since
the only variant in the device node is the UUID as per the crypttab
changes. This makes easier the connection of fstab entries and crypttab
entries.