Fix test for allowing the installation source to be on the root fs (#462769).
This goes back to how the stage2 stuff used to work.  By this point in the
upgrade process, the installation source would have already been mounted so
trying to mount it again to test for /etc/redhat-release would result in
an error from mount.  However, we no longer have the source mounted when
performing this test so we need to make sure to try mounting the protected
partitions just like every other.