Don't worry about errors looking up protected partitions on upgrades.
This mostly affects preupgrade, where the repo (and therefore what's passed
in as the methodstr) can be a UUID for a logical volume.  At this point,
lvm is not activated so looking up the device matching the UUID will of course
fail.  That's not really an error on the upgrade case since we won't be
formatting any partitions.  On the installation case, this is an error.  And
keeping it an error here means we continue not allowing HDISO installs from
an LV or RAID volume.