Fix lang-name generation + fix traceback with LANG=C
Generating lang-name depended on old behavior of rhpl.translate for looking
in po/ for the translation files.  The regular gettext module doesn't do this
which means we don't get translated language names unless you already
had anaconda installed.  So use gettext.GNUTranslation to directly open
the .mo file

For LANG=C, we need to ensure that the output charset is set to UTF-8