Fix mk-s390-cdboot on s390x (#184648)
Related to #184648.  The C program provided by IBM was not
working.  In buildinstall.log from distill, we saw that it
was core dumping.  I've patched the source to not use fixed
buffer sizes for command line arguments, which seemed to be
the cause of the problem.  I also added some error checking
and reporting in case it completely fails.