Work around gtk2 bug rh465541
gtk2-2.14.2 and newer has a bug where the return value of
gtk.gdk.get_width() does not change after gtk has been initialized, as we
call xrandr to change the resolution this is a problem.

This patch works around this by not initializing gtk before calling xrandr,
this means that we will call xrandr -s even if the resolution already is
what we want (usually 800x600) but chances of that being the case are small
anyways.