Do not rely on loaderData->noDns to tell if we have DNS configured.
Since we're using NetworkManager, just try things and see if they
work.  For example, try to do a hostname lookup when mounting the
NFS source.  If it works, we have working DNS and can move on.
If it fails, we can tell the user they are missing DNS information
and maybe they'll get the hint to either configure DNS servers or
to specify the NFS source using an IP address.