Mark iscsi disks not used for / as autostart (rh461840)
What is happening here is that with / on iscsi we do not want to try and start
the node(s) corresponding to the raid / logvol / partition used to autostart
because they are already started and the node database may even contain wrong
info causing problems as described in bug 437891. However for /somedir on iscsi
we do want to set the node to autostart, as there the node will not have been
started from the initrd, so without starting it we won't have a connection to
the disk for /somedir.

Thus this commit sets node.startup to automatic for non / containing iscsi
nodes only.