Write NM_CONTROLLED=yes rather than NM_CONTROLLED=
I was told by upstream that NM_CONTROLLED= should be sufficient
for nm-system-settings to pick up an interface for control.  But
the parsing in the ifcfg-fedora plugin for nm-system-settings is a
little strange (shvar.c) and what happens when NM_CONTROLLED is
empty isn't entirely clear.  Set it to yes to avoid any possible
problems.