Enable CCW devices used for installation (#253075)
IBM had requested that on s390x we add the cio_ignore=all parameter
to the default boot file.  Well, cio_ignore all except for the 3270
terminal.  The motivation for this change was to decrease device
discovery time on systems with thousands of subchannels.

The fix was only part of the solution.  When you use cio_ignore to
disable everything, you need to explicitly free the device from
cio_ignore and then enable it using the magic of /sbin/chccwdev.
This patch adds the un-ignore capability to linuxrc.s390.