Say we are unable to configure the network interface (#467960)
In stage 2 when you get the 'Enable network interface' dialog, if
you select an interface that is not able to connect to a network,
we just wait and then take the user back to the Enable network
interface dialog.  Before going back to the dialog, tell the user
we were unable to configure the selected interface.