Catch UnicodeDecodeError so traceback messages display anyway.
I've seen this happen in recent nightly builds.  UnicodeDecodeError
gets thrown by unicode() and the traceback window doesn't finish
rendering and we get two tracebacks dumped to the console, one for
the traceback window tracebacking and the original one we were
going to display.

This workaround at least gets us traceback windows displaying.
I don't know enough about unicode() to do anything more with it
right now, so this seems reasonable.

If the exception is thrown, I log a message so we have some record
of the unicode failure.

My guess is that the i18n files are invalid and/or incompatible
with what we're trying to shove them through in Python.  But
that's just a guess.