Don't treat encrypted PVs as available if we don't have the key. (#465240)
This arises from the fact that we determine whether or not a device
is a PV by looking at the partition type instead of probing the
device's contents like with other filesystems/ptypes. So, if we have
a PV that the user chose not to provide a passphrase for, we should
not include it in the list of available PVs.