Make sure the productVersion given by .treeinfo exists in bugzilla (#468657).
Prerelease version of Fedora sometimes get version numbers in .treeinfo or
.buildstamp like 10-Beta.  However, this version does not exist in bugzilla so
we get an error trying to save the bug.  This could also happen in respins.
So, grab a list of all the versions that are valid for the product out of
bugzilla and make sure productVersion exists.  If not, either use the
develVersion from the installclass or just use the last one out of the the
list.