Call startNewt earlier than network bring up (#469171).
This needs to be called before newtWinMessage, which can happen in various
conditions in the networking code paths.  So we can either create our own
newtWinMessage to call startNewt for us (preferred fix for the future), add
more calls to startNewt throughout the networking code, or just move the first
call earlier.  For F10, the last seems like the most straightforward fix.