Make sure we look up the IP address for the correct device (#469439)
The VNC launch code in vnc.py needed an update to work better with
NetworkManager.  When collecting the hostname and IP address, it was
assuming the first device in the netdevices list is our active NIC,
which may or may not be true.

Added getActiveNetDevs() in network.py to ask NetworkManager for a
list of all currently configured interfaces.  Return a list of
device names.  A list seems a bit pointless, but I'd like to have
this in place now for future improvements where we might need to
handle more than one active NIC during installation.  After all,
NM can do that.

The message reported by anaconda once VNC is ready will contain the
FQDN:DISPLAY_NUMBER (IP ADDRESS), if it can.  If it can't find your
IP address, it leaves that out.  If it can't find your hostname, it
also leaves that out.