Let users edit net settings on network failure in stage 1 (#465887)
This one has been around for a while.  If a user selects the wrong
network interface in loader and waits for the failure message to
appear, they are taken back to the 'select an interface' screen.
They select the right interface, but loader used the TCP/IP settings
entered the first time around.

On network configuration failure and if we are doing an interactive
install, blank out ipv4method and ipv6method so that we take the
user to the TCP/IP configuration screen.  I use ipv4method and
ipv6method to figure out if a kickstart installation is beginning,
so by blanking them out for the interactive install case, we get to
take the user back to the TCP/IP settings screen.