Prevent traceback when there are no network devices (#469339)
If a user boots from boot.iso, the installer will bring up the net
config dialog to configure the network interface so packages can
be found and installed.  If a user has no available network devices,
the window appears but nothing is in the combo box.  If they click
the combo box, you get a traceback.

This patch returns False from all enableNetwork() functions so the
user sees the 'No Network Available' dialog and is given the Exit
Installer button.