Do not stop and restart iscsid when rescanning disks/partitions (#470223)
anaconda kills and restarts iscsid each time an iscsi / zfcp disk gets added,
resulting in messages like these on tty4:

<28> Nov  6 11:30:06 iscsid: connection1:0 is operational now
<28> Nov  6 11:30:07 iscsid: iscsid shutting down.
<28> Nov  6 11:30:08 iscsid: iscsid with pid=1359 started
<28> Nov  6 11:30:09 iscsid: session
[iqn.2005-03.com.max:testdisk,192.168.1.100:3260] already running
<28> Nov  6 11:30:09 iscsid: connection1:0 is operational after recovery (1
attempts)

Esp. the "connection1:0 is operational after recovery (1 attempts)" clearly
indicates that this is not a good thing todo. iscsid is designed as a kernel
helper daemon, which when iscsi is used should be started and never stopped
again.

Note that this means iscsi.startup() also will not get called during the
partitionobjinit step, this is a no-op the first time as then the initiator
name has not been set yet, and if we go back and forth over this step we do not
want to start / stop iscsid, just as we don't want to do that for each iscsi
target added.