kickstart, iscsi do not call iscsi.startup after startIBFT has been called
The current kickstart IscsiName code calls iscsi.startup after calling
iscsi.startIBFT, but startIBFT will call iscsi.addTarget when an IBFT target
is found, which will in turn then call iscsi.startup. iscsi.startup is not
protected against being called twice, and currently doing so happens to not
have adverse side effects, but we reallt should not be relying on that.

Note: when no IBFT target is found there is no reason to call iscsi.startup