Prevent traceback for vnc installs on KVM guests (#470559)
If you are installing in a KVM guest and pass 'vnc' on the boot line,
getDefaultHostname() in network.py will traceback because anaconda.id
contains nothing in this particular use case.  Wrap the offending
line in try/except and fall back on None if we can't pick the hostname
from anaconda.id.network.

For KVM installs, this causes vnc.py to report 'Please connect your
vnc client to IPADDRESS:DISPLAY to begin the install', which is good
enough for me.