Add comps groups for new repos that are added (#470653)
For new repos that are added, we should also be sure to add their comps
information to the base comps object.  This largely cut and pastes bits from
yum's doGroupSetup(), but that can't be done on a per-repo basis.  We should
add some API, but this will do for now