Allow installing grub on the MBR if /boot is on mdraid (#217176)
This patch removes a special case when /boot (or / without a separate /boot)
is on mdraid. For some reason we only allow installing grub to the mdX device
then instead of allowing both the mbr of the 1st disk and the device holding
/boot.

I've searched through the anaconda history and this special handling of
mdraid /boot has been there since before Feb. 2002, it was inherited from
before booty was introduced and it does not seem to make sense. Either we
cannot have /boot on mdraid at all (in all cases except for raid1) or we can
have /boot on mdraid (in case of raid1) and then it does not matter if the
bootstrap of grub starts on the mbr or on the first sector of a partition.

This fixes 217176