Better naming for LVM volume groups and logical volumes (#461682)
Try to name volume groups as vg_HOSTNAME and logical volumes as
lv_MOUNTPOINT, if we can.  Swap partitions will be lv_swapNN where
NN is a unique number in the instance where more than one swap
partition in use.  The / partition will get the name lv_root.

Fall back on the old naming system (VolGroupNN for volume groups
and LogVolNN for logical volumes) for people doing custom setup
or where the hostname is localhost.

For swap partition naming, tack on an NN designation when there
are more than 1 swap partitions requested.  If only one is
requested, it will be "lv_swap".