Send unique vendor class identifier unless user specifies one.
Restores functionality we had before the NM changeover.  This was
originally requested by gurulabs.com.  Instead of sending just
'anaconda' as the vendor class identifier, send something like:

    anaconda-Linux 2.6.27.5-117.fc10.i686 i686

Of course, if we fail to collect that information, fall back on
'anaconda'.  Guru Labs wanted this functionality for PXE booting
as it allowed them to more easily control per-client/per-release
settings from the DHCP server.  My fault for not adding this back
when the NM changeover happened.