Prevent network install when no network devices are found (#470144)
A corner case, but if a user boots up with boot.iso and has no
network devices available, we get all the way to package selection
and user is told they cannot continue.  This is after all of the
other screens and partitioning.

In the welcome screen, perform a check to see if the user has booted
the installer to a point where it requires a network install for the
packages.  If so, tell the user they cannot continue and exit the
installer.