Preserve existing network configuration files during install (#461550)
The suggestion for this patch came from OLPC (or someone with an
@laptop.org email address).  The user was doing a kickstart install
and adding a local yum repo that provided an RPM of site-specific
configuration files, including ifcfg-DEVICE files.  This package
would get installed just fine, but then anaconda drives over the
configuration files after package installation.

If we see any configuration file on the target system during
installation, skip over it and don't write a new one.  This
should handle upgrades and kickstart install types like OLPC was
describing.