Use stacks instead of tracebacks in traceback handlers.
This allows us to generate traceback-like stacks on demand
using inspect.stack(). stack() returns a list of tuples
representing the stack frames at the time of invocation. A
traceback object can be converted to a list of these
tuples by calling inspect.getinnerframes(tb). NOTE: the
stack() function returns the frames in the opposite order
of a traceback's.