mdraid1: default to putting grub on partition instead of mbr (#217176)
As previously discussed we do want to allow installing the bootloader on the
MBR in the raid 1 case (which until some time ago we did not allow) (#217176),
but we do not want to change our current default of installing to the partition
(which gives us the mirror redundancy, due to special handling in booty).

This patch effectively reverts commit 48d582347492493c87b2b368394fd8e7342688bc,
which removed the raid1 case special handling, and then adds mbr support to the
raid special case in fsset.py, and add some special case handling to
bootloader.py to not default to the mbr in this case.