Only write the short hostname to the localhost line (#474086)
Do not write the FQDN to the localhost line, only write the short
hostname if it's something other than 'localhost'.  Only write the
FQDN line to /etc/hosts if we have an IP address for the host.

Without this patch, users performing media installs who enter an
FQDN in the hostname field end up with their FQDN == 127.0.0.1,
which messes up things, such as sendmail.  Writing the short name
is what we really want here.

[rawhide]